sports-llm/requirements.txt
# Claude-Only Architecture - Minimal Dependencies

# Vector Search (lightweight embedding model only)
sentence-transformers>=2.2.0

# MongoDB
pymongo>=4.6.0
motor>=3.3.0  # Async MongoDB driver

# FastAPI (for driver.py)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests for Claude API
aiohttp>=3.9.0

# Minimal data processing
numpy>=1.24.0

